Update the style of headings
This is to better fit the HTML rendered version by jekyll
Also remember to double escape the slashes for the latex in the markdown
code so they reach to the HTML and can then benefit from MathJax. Added
note at the end.

Update SCSS file to style the headings in the CV